chore(schema): move legacy utils to a dedicated module
